Add files via upload
The script comes with an example right at the end. It's really easy to use, just feed the findGraphIntersections function a dataframe with variables x and y (x coordinates of each point, and y coordinates of each point). The algorithm will check for intersections using a system of linear equations with the help of the matlib library (required to run the script)  and return them. If you set output as TRUE, then it will draw the graph and highlight the intersection as well.
Example CSV file is included. Note that this assumes a closed-loop TSP, but please go ahead and change it if you like.