Add files via upload
Example graphs, with some example code given at the end of the script. Just make sure to load all the function in the R script (just select the entire script and run it).
The Monte Carlo Method helps find the area of irregular shapes with crossings. I've added to example graphs, one which is fairly circular (ratio of area of shape to total area is almost 0.5), and a slim graph (ratio is approximately 0.38). The more points you add, the closer you will get to the 'truth', however be prepared for a long wait...